+++ mayan/apps/documents/tasks.py
@@ -265,26 +265,36 @@ def task_document_version_export(
 # Trash can
 
 @app.task(ignore_result=True)
-def task_trash_can_empty():
+def task_trash_can_empty(user_id=None):
     TrashedDocument = apps.get_model(
         app_label='documents', model_name='TrashedDocument'
     )
 
     for trashed_document in TrashedDocument.objects.all():
         task_trashed_document_delete.apply_async(
-            kwargs={'trashed_document_id': trashed_document.pk}
+            kwargs={
+                'trashed_document_id': trashed_document.pk,
+                'user_id': user_id
+            }
         )
 
 
 # Trashed document
 
 @app.task(ignore_result=True)
-def task_trashed_document_delete(trashed_document_id):
+def task_trashed_document_delete(trashed_document_id, user_id=None):
     TrashedDocument = apps.get_model(
         app_label='documents', model_name='TrashedDocument'
     )
+    User = get_user_model()
+
+    if user_id:
+        user = User.objects.get(pk=user_id)
+    else:
+        user = None
 
     logger.debug(msg='Executing')
     trashed_document = TrashedDocument.objects.get(pk=trashed_document_id)
+    trashed_document._event_actor = user
     trashed_document.delete()
     logger.debug(msg='Finished')