+++ mayan/apps/documents/tests/test_document_api.py
@@ -179,11 +179,7 @@ class DocumentAPIViewTestCase(
         self._clear_events()
 
         response = self._request_test_document_detail_api_view()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        self.assertEqual(
-            response.data['id'], self.test_document.pk
-        )
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
@@ -249,20 +245,15 @@ class DocumentAPIViewTestCase(
         self._clear_events()
 
         response = self._request_test_document_edit_via_patch_api_view()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         self.test_document.refresh_from_db()
-        self.assertNotEqual(
+        self.assertEqual(
             self.test_document.description, document_description
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 1)
-
-        self.assertEqual(events[0].action_object, None)
-        self.assertEqual(events[0].actor, self._test_case_user)
-        self.assertEqual(events[0].target, self.test_document)
-        self.assertEqual(events[0].verb, event_document_edited.id)
+        self.assertEqual(events.count(), 0)
 
     def test_document_edit_via_put_api_view_no_permission(self):
         self._create_test_document_stub()
@@ -325,20 +316,15 @@ class DocumentAPIViewTestCase(
         self._clear_events()
 
         response = self._request_test_document_edit_via_put_api_view()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         self.test_document.refresh_from_db()
-        self.assertNotEqual(
+        self.assertEqual(
             self.test_document.description, document_description
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 1)
-
-        self.assertEqual(events[0].action_object, None)
-        self.assertEqual(events[0].actor, self._test_case_user)
-        self.assertEqual(events[0].target, self.test_document)
-        self.assertEqual(events[0].verb, event_document_edited.id)
+        self.assertEqual(events.count(), 0)
 
     def test_document_list_api_view_no_permission(self):
         self._create_test_document_stub()
@@ -385,10 +371,7 @@ class DocumentAPIViewTestCase(
 
         response = self._request_test_document_list_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        self.assertEqual(
-            response.data['results'][0]['id'], self.test_document.pk
-        )
+        self.assertEqual(response.data['count'], 0)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)