+++ mayan/apps/documents/tests/test_favorite_document_api.py
@@ -22,7 +22,7 @@ class FavoriteDocumentAPIViewTestCase(
         self._create_test_document_stub()
 
     def test_favorite_document_create_api_view_no_permission(self):
-        favorite_document_count = FavoriteDocument.objects.count()
+        favorite_document_count = FavoriteDocument.valid.count()
 
         self._clear_events()
 
@@ -31,13 +31,13 @@ class FavoriteDocumentAPIViewTestCase(
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         self.assertEqual(
-            FavoriteDocument.objects.count(), favorite_document_count
+            FavoriteDocument.valid.count(), favorite_document_count
         )
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
     def test_favorite_document_create_api_view_with_access(self):
-        favorite_document_count = FavoriteDocument.objects.count()
+        favorite_document_count = FavoriteDocument.valid.count()
 
         self.grant_access(
             obj=self.test_document,
@@ -50,7 +50,7 @@ class FavoriteDocumentAPIViewTestCase(
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
 
         self.assertEqual(
-            FavoriteDocument.objects.count(), favorite_document_count + 1
+            FavoriteDocument.valid.count(), favorite_document_count + 1
         )
         self.assertEqual(
             self.test_favorite_document.document.pk, self.test_document.pk
@@ -60,7 +60,7 @@ class FavoriteDocumentAPIViewTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_trashed_document_favorite_document_create_api_view_with_access(self):
-        favorite_document_count = FavoriteDocument.objects.count()
+        favorite_document_count = FavoriteDocument.valid.count()
 
         self.grant_access(
             obj=self.test_document,
@@ -72,13 +72,10 @@ class FavoriteDocumentAPIViewTestCase(
         self._clear_events()
 
         response = self._request_test_favorite_document_create_api_view()
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         self.assertEqual(
-            FavoriteDocument.objects.count(), favorite_document_count + 1
-        )
-        self.assertEqual(
-            self.test_favorite_document.document.pk, self.test_document.pk
+            FavoriteDocument.valid.count(), favorite_document_count
         )
 
         events = self._get_test_events()
@@ -87,7 +84,7 @@ class FavoriteDocumentAPIViewTestCase(
     def test_favorite_document_delete_api_view_no_permission(self):
         self._test_document_favorite_add()
 
-        favorite_document_count = FavoriteDocument.objects.count()
+        favorite_document_count = FavoriteDocument.valid.count()
 
         self._clear_events()
 
@@ -95,7 +92,7 @@ class FavoriteDocumentAPIViewTestCase(
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         self.assertEqual(
-            FavoriteDocument.objects.count(), favorite_document_count
+            FavoriteDocument.valid.count(), favorite_document_count
         )
 
         events = self._get_test_events()
@@ -104,7 +101,7 @@ class FavoriteDocumentAPIViewTestCase(
     def test_favorite_document_delete_api_view_with_access(self):
         self._test_document_favorite_add()
 
-        favorite_document_count = FavoriteDocument.objects.count()
+        favorite_document_count = FavoriteDocument.valid.count()
 
         self.grant_access(
             obj=self.test_document,
@@ -117,7 +114,7 @@ class FavoriteDocumentAPIViewTestCase(
         self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
 
         self.assertEqual(
-            FavoriteDocument.objects.count(), favorite_document_count - 1
+            FavoriteDocument.valid.count(), favorite_document_count - 1
         )
 
         events = self._get_test_events()
@@ -126,7 +123,7 @@ class FavoriteDocumentAPIViewTestCase(
     def test_trashed_document_favorite_document_delete_api_view_with_access(self):
         self._test_document_favorite_add()
 
-        favorite_document_count = FavoriteDocument.objects.count()
+        favorite_document_count = FavoriteDocument.valid.count()
 
         self.grant_access(
             obj=self.test_document,
@@ -138,10 +135,10 @@ class FavoriteDocumentAPIViewTestCase(
         self._clear_events()
 
         response = self._request_test_favorite_document_delete_api_view()
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         self.assertEqual(
-            FavoriteDocument.objects.count(), favorite_document_count - 1
+            FavoriteDocument.valid.count(), favorite_document_count - 1
         )
 
         events = self._get_test_events()
@@ -191,11 +188,7 @@ class FavoriteDocumentAPIViewTestCase(
         self._clear_events()
 
         response = self._request_test_favorite_document_detail_api_view()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            response.data['document']['label'],
-            self.test_document.label
-        )
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
@@ -203,7 +196,7 @@ class FavoriteDocumentAPIViewTestCase(
     def test_favorite_document_list_api_view_no_permission(self):
         self._test_document_favorite_add()
 
-        favorite_document_count = FavoriteDocument.objects.count()
+        favorite_document_count = FavoriteDocument.valid.count()
 
         self._clear_events()
 
@@ -217,7 +210,7 @@ class FavoriteDocumentAPIViewTestCase(
     def test_favorite_document_list_api_view_with_access(self):
         self._test_document_favorite_add()
 
-        favorite_document_count = FavoriteDocument.objects.count()
+        favorite_document_count = FavoriteDocument.valid.count()
 
         self.grant_access(
             obj=self.test_document,
@@ -240,7 +233,7 @@ class FavoriteDocumentAPIViewTestCase(
     def test_trashed_document_favorite_document_list_api_view_with_access(self):
         self._test_document_favorite_add()
 
-        favorite_document_count = FavoriteDocument.objects.count()
+        favorite_document_count = FavoriteDocument.valid.count()
 
         self.grant_access(
             obj=self.test_document,
@@ -253,11 +246,7 @@ class FavoriteDocumentAPIViewTestCase(
 
         response = self._request_test_favorite_document_list_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data['count'], favorite_document_count)
-        self.assertEqual(
-            response.data['results'][0]['document']['label'],
-            self.test_document.label
-        )
+        self.assertEqual(response.data['count'], favorite_document_count - 1)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)