+++ mayan/apps/documents/tests/test_favorite_document_models.py
@@ -7,7 +7,7 @@ from .mixins.favorite_document_mixins import FavoriteDocumentTestMixin
 
 
 @override_settings(DOCUMENTS_FAVORITE_COUNT=2)
-class TrashedDocumentTestCase(
+class FavoriteDocumentModelTestCase(
     FavoriteDocumentTestMixin, GenericDocumentTestCase
 ):
     auto_upload_test_document = False
@@ -25,5 +25,14 @@ class TrashedDocumentTestCase(
         self._test_document_favorite_add()
 
         self.assertFalse(
-            FavoriteDocument.objects.filter(document=first_favorite_document).exists()
+            FavoriteDocument.valid.filter(document=first_favorite_document).exists()
+        )
+
+    def test_trashed_document_favorite_document_add(self):
+        self._create_test_document_stub()
+        self.test_document.delete()
+        self._test_document_favorite_add()
+
+        self.assertFalse(
+            FavoriteDocument.valid.filter(document=self.test_document).exists()
         )