+++ mayan/apps/documents/tests/test_recently_accessed_document_api.py
@@ -2,7 +2,6 @@ from rest_framework import status
 
 from mayan.apps.rest_api.tests.base import BaseAPITestCase
 
-from ..models.recently_accessed_document_models import RecentlyAccessedDocument
 from ..permissions import permission_document_view
 
 from .mixins.document_mixins import DocumentTestMixin
@@ -22,25 +21,16 @@ class RecentlyAccessedDocumentAPIViewTestCase(
         self._create_test_document_stub()
 
     def test_recently_accessed_document_api_list_view_no_activity(self):
-        recently_accessed_document_count = RecentlyAccessedDocument.objects.count()
-
         self._clear_events()
 
         response = self._request_recently_accessed_document_list_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data['count'], 0)
 
-        self.assertEqual(
-            RecentlyAccessedDocument.objects.count(),
-            recently_accessed_document_count
-        )
-
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
     def test_recently_accessed_document_api_list_view_with_activity_no_permission(self):
-        recently_accessed_document_count = RecentlyAccessedDocument.objects.count()
-
         self.test_document.add_as_recent_document_for_user(
             user=self._test_case_user
         )
@@ -51,17 +41,10 @@ class RecentlyAccessedDocumentAPIViewTestCase(
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data['count'], 0)
 
-        self.assertEqual(
-            RecentlyAccessedDocument.objects.count(),
-            recently_accessed_document_count + 1
-        )
-
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
     def test_recently_accessed_document_api_list_view_with_activity_with_access(self):
-        recently_accessed_document_count = RecentlyAccessedDocument.objects.count()
-
         self.test_document.add_as_recent_document_for_user(
             user=self._test_case_user
         )
@@ -75,20 +58,14 @@ class RecentlyAccessedDocumentAPIViewTestCase(
         response = self._request_recently_accessed_document_list_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(
-            response.data['results'][0]['document']['id'], self.test_document.pk
-        )
-
-        self.assertEqual(
-            RecentlyAccessedDocument.objects.count(),
-            recently_accessed_document_count + 1
+            response.data['results'][0]['document']['id'],
+            self.test_document.pk
         )
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
     def test_trashed_document_recently_accessed_document_api_list_view_with_activity_with_access(self):
-        recently_accessed_document_count = RecentlyAccessedDocument.objects.count()
-
         self.test_document.add_as_recent_document_for_user(
             user=self._test_case_user
         )
@@ -103,14 +80,7 @@ class RecentlyAccessedDocumentAPIViewTestCase(
 
         response = self._request_recently_accessed_document_list_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            response.data['results'][0]['document']['id'], self.test_document.pk
-        )
-
-        self.assertEqual(
-            RecentlyAccessedDocument.objects.count(),
-            recently_accessed_document_count + 1
-        )
+        self.assertEqual(response.data['count'], 0)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)