+++ mayan/apps/documents/tests/test_recently_accessed_document_views.py
@@ -18,10 +18,15 @@ class RecentlyAccessedDocumentViewTestCase(
             user=self._test_case_user
         )
 
+        self._clear_events()
+
         response = self._request_test_recently_accessed_document_list_view()
         self.assertNotContains(
             response=response, text=self.test_document.label, status_code=200
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_recently_accessed_document_list_view_with_access(self):
         self.test_document.add_as_recent_document_for_user(
@@ -32,10 +37,15 @@ class RecentlyAccessedDocumentViewTestCase(
             obj=self.test_document, permission=permission_document_view
         )
 
+        self._clear_events()
+
         response = self._request_test_recently_accessed_document_list_view()
         self.assertContains(
             response=response, text=self.test_document.label, status_code=200
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_trashed_recently_accessed_document_list_view_with_access(self):
         self.test_document.add_as_recent_document_for_user(
@@ -47,7 +57,12 @@ class RecentlyAccessedDocumentViewTestCase(
             obj=self.test_document, permission=permission_document_view
         )
 
+        self._clear_events()
+
         response = self._request_test_recently_accessed_document_list_view()
         self.assertNotContains(
             response=response, text=self.test_document.label, status_code=200
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)