+++ mayan/apps/documents/tests/test_recently_created_document_views.py
@@ -14,20 +14,30 @@ class RecentlyCreatedDocumentViewTestCase(
         self._create_test_document_stub()
 
     def test_recently_created_document_list_view_no_permission(self):
+        self._clear_events()
+
         response = self._request_test_recently_created_document_list_view()
         self.assertNotContains(
             response=response, text=self.test_document.label, status_code=200
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_recently_created_document_list_view_with_access(self):
         self.grant_access(
             obj=self.test_document, permission=permission_document_view
         )
 
+        self._clear_events()
+
         response = self._request_test_recently_created_document_list_view()
         self.assertContains(
             response=response, text=self.test_document.label, status_code=200
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_trashed_recently_created_document_list_view_with_access(self):
         self.grant_access(
@@ -35,7 +45,12 @@ class RecentlyCreatedDocumentViewTestCase(
         )
         self.test_document.delete()
 
+        self._clear_events()
+
         response = self._request_test_recently_created_document_list_view()
         self.assertNotContains(
             response=response, text=self.test_document.label, status_code=200
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)