+++ mayan/apps/documents/tests/test_trashed_document_api.py
@@ -68,6 +68,27 @@ class TrashedDocumentAPIViewTestCase(
         self.assertEqual(events[0].target, self.test_document)
         self.assertEqual(events[0].verb, event_document_trashed.id)
 
+    def test_trashed_document_trash_api_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_document_trash
+        )
+
+        self.test_document.delete()
+
+        trashed_document_count = Document.trash.count()
+        document_count = Document.valid.count()
+
+        self._clear_events()
+
+        response = self._request_test_document_trash_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertEqual(Document.trash.count(), trashed_document_count)
+        self.assertEqual(Document.valid.count(), document_count)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
     def test_trashed_document_delete_api_view_no_permission(self):
         self.test_document.delete()
 
@@ -98,7 +119,7 @@ class TrashedDocumentAPIViewTestCase(
         self._clear_events()
 
         response = self._request_test_trashed_document_delete_api_view()
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        self.assertEqual(response.status_code, status.HTTP_202_ACCEPTED)
 
         self.assertEqual(Document.trash.count(), trashed_document_count - 1)
         self.assertEqual(Document.valid.count(), document_count)
@@ -110,6 +131,26 @@ class TrashedDocumentAPIViewTestCase(
         self.assertEqual(events[0].actor, self._test_case_user)
         self.assertEqual(events[0].target, self.test_document_type)
         self.assertEqual(events[0].verb, event_trashed_document_deleted.id)
+
+    def test_non_trashed_document_delete_api_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document,
+            permission=permission_trashed_document_delete
+        )
+
+        trashed_document_count = Document.trash.count()
+        document_count = Document.valid.count()
+
+        self._clear_events()
+
+        response = self._request_test_trashed_document_delete_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertEqual(Document.trash.count(), trashed_document_count)
+        self.assertEqual(Document.valid.count(), document_count)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_trashed_document_detail_api_view_no_permission(self):
         self.test_document.delete()
@@ -152,32 +193,21 @@ class TrashedDocumentAPIViewTestCase(
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
-
-class TrashedDocumentImageAPIViewTestCase(
-    TrashedDocumentAPIViewTestMixin, DocumentTestMixin, BaseAPITestCase
-):
-    def test_trashed_document_image_api_view_no_permission(self):
-        self.test_document.delete()
-
-        self._clear_events()
-
-        response = self._request_test_trashed_document_image_api_view()
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-        events = self._get_test_events()
-        self.assertEqual(events.count(), 0)
-
-    def test_trashed_document_image_api_view_with_access(self):
-        self.test_document.delete()
+    def test_non_trashed_document_detail_api_view_with_access(self):
         self.grant_access(
-            obj=self.test_document,
-            permission=permission_document_version_view
+            obj=self.test_document, permission=permission_document_view
         )
 
+        trashed_document_count = Document.trash.count()
+        document_count = Document.valid.count()
+
         self._clear_events()
 
-        response = self._request_test_trashed_document_image_api_view()
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        response = self._request_test_trashed_document_detail_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertEqual(Document.trash.count(), trashed_document_count)
+        self.assertEqual(Document.valid.count(), document_count)
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
@@ -212,6 +242,20 @@ class TrashedDocumentImageAPIViewTestCase(
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
+    def test_non_trashed_document_list_api_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_document_view
+        )
+
+        self._clear_events()
+
+        response = self._request_test_trashed_document_list_api_view()
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data['count'], 0)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
     def test_trashed_document_restore_via_get_api_view_no_permission(self):
         self.test_document.delete()
 
@@ -232,6 +276,26 @@ class TrashedDocumentImageAPIViewTestCase(
     def test_trashed_document_restore_via_get_api_view_with_access(self):
         self.test_document.delete()
 
+        self.grant_access(
+            obj=self.test_document,
+            permission=permission_trashed_document_restore
+        )
+
+        trashed_document_count = Document.trash.count()
+        document_count = Document.valid.count()
+
+        self._clear_events()
+
+        response = self._request_test_trashed_document_restore_via_get_api_view()
+        self.assertEqual(response.status_code, status.HTTP_405_METHOD_NOT_ALLOWED)
+
+        self.assertEqual(Document.trash.count(), trashed_document_count)
+        self.assertEqual(Document.valid.count(), document_count)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_non_trashed_document_restore_via_get_api_view_with_access(self):
         self.grant_access(
             obj=self.test_document,
             permission=permission_trashed_document_restore
@@ -294,3 +358,67 @@ class TrashedDocumentImageAPIViewTestCase(
         self.assertEqual(events[0].actor, self._test_case_user)
         self.assertEqual(events[0].target, self.test_document)
         self.assertEqual(events[0].verb, event_trashed_document_restored.id)
+
+    def test_non_trashed_document_restore_via_post_api_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document,
+            permission=permission_trashed_document_restore
+        )
+
+        trashed_document_count = Document.trash.count()
+        document_count = Document.valid.count()
+
+        self._clear_events()
+
+        response = self._request_test_trashed_document_restore_via_post_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertEqual(Document.trash.count(), trashed_document_count)
+        self.assertEqual(Document.valid.count(), document_count)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+
+class TrashedDocumentImageAPIViewTestCase(
+    TrashedDocumentAPIViewTestMixin, DocumentTestMixin, BaseAPITestCase
+):
+    def test_trashed_document_image_api_view_no_permission(self):
+        self.test_document.delete()
+
+        self._clear_events()
+
+        response = self._request_test_trashed_document_image_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_trashed_document_image_api_view_with_access(self):
+        self.test_document.delete()
+        self.grant_access(
+            obj=self.test_document,
+            permission=permission_document_version_view
+        )
+
+        self._clear_events()
+
+        response = self._request_test_trashed_document_image_api_view()
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_non_trashed_document_image_api_view_with_access(self):
+        self.grant_access(
+            obj=self.test_document,
+            permission=permission_document_version_view
+        )
+
+        self._clear_events()
+
+        response = self._request_test_trashed_document_image_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)