+++ mayan/apps/documents/tests/test_trashed_document_views.py
@@ -1,3 +1,7 @@
+from ..events import (
+    event_document_trashed, event_trashed_document_deleted,
+    event_trashed_document_restored
+)
 from ..models.document_models import Document
 from ..models.trashed_document_models import TrashedDocument
 from ..permissions import (
@@ -23,6 +27,8 @@ class DocumentTrashViewTestCase(
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_document_trash_get_view()
         self.assertEqual(response.status_code, 404)
 
@@ -30,6 +36,9 @@ class DocumentTrashViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_document_trash_get_view_with_access(self):
         self.grant_access(
@@ -39,6 +48,8 @@ class DocumentTrashViewTestCase(
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_document_trash_get_view()
         self.assertEqual(response.status_code, 200)
 
@@ -46,6 +57,9 @@ class DocumentTrashViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_trashed_document_trash_get_view_with_access(self):
         self.grant_access(
@@ -57,6 +71,8 @@ class DocumentTrashViewTestCase(
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_document_trash_get_view()
         self.assertEqual(response.status_code, 404)
 
@@ -65,9 +81,14 @@ class DocumentTrashViewTestCase(
             TrashedDocument.objects.count(), trashed_document_count
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
     def test_document_trash_post_view_no_permission(self):
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
+
+        self._clear_events()
 
         response = self._request_test_document_trash_post_view()
         self.assertEqual(response.status_code, 404)
@@ -76,6 +97,9 @@ class DocumentTrashViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_document_trash_post_view_with_access(self):
         self.grant_access(
@@ -85,6 +109,8 @@ class DocumentTrashViewTestCase(
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_document_trash_post_view()
         self.assertEqual(response.status_code, 302)
 
@@ -92,6 +118,14 @@ class DocumentTrashViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count + 1
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_document)
+        self.assertEqual(events[0].verb, event_document_trashed.id)
 
     def test_trashed_document_trash_post_view_with_access(self):
         self.grant_access(
@@ -103,6 +137,8 @@ class DocumentTrashViewTestCase(
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_document_trash_post_view()
         self.assertEqual(response.status_code, 404)
 
@@ -110,6 +146,9 @@ class DocumentTrashViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
 
 class TrashedDocumentViewTestCase(
@@ -127,6 +166,8 @@ class TrashedDocumentViewTestCase(
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_trashed_document_delete_get_view()
         self.assertEqual(response.status_code, 404)
 
@@ -136,6 +177,9 @@ class TrashedDocumentViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_trashed_document_delete_get_view_with_access(self):
         self.test_document.delete()
@@ -147,6 +191,8 @@ class TrashedDocumentViewTestCase(
             permission=permission_trashed_document_delete
         )
 
+        self._clear_events()
+
         response = self._request_test_trashed_document_delete_get_view()
         self.assertEqual(response.status_code, 200)
 
@@ -157,10 +203,15 @@ class TrashedDocumentViewTestCase(
             TrashedDocument.objects.count(), trashed_document_count
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
     def test_trashed_document_delete_post_view_no_permission(self):
         self.test_document.delete()
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
+
+        self._clear_events()
 
         response = self._request_test_trashed_document_delete_post_view()
         self.assertEqual(response.status_code, 404)
@@ -170,14 +221,20 @@ class TrashedDocumentViewTestCase(
             TrashedDocument.objects.count(), trashed_document_count
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
     def test_trashed_document_delete_post_view_with_access(self):
         self.test_document.delete()
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
 
         self.grant_access(
-            obj=self.test_document, permission=permission_trashed_document_delete
+            obj=self.test_document,
+            permission=permission_trashed_document_delete
         )
+
+        self._clear_events()
 
         response = self._request_test_trashed_document_delete_post_view()
         self.assertEqual(response.status_code, 302)
@@ -187,13 +244,26 @@ class TrashedDocumentViewTestCase(
             TrashedDocument.objects.count(), trashed_document_count - 1
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_document_type)
+        self.assertEqual(events[0].verb, event_trashed_document_deleted.id)
+
     def test_trashed_document_list_view_no_permission(self):
         self.test_document.delete()
+
+        self._clear_events()
 
         response = self._request_test_trashed_document_list_view()
         self.assertNotContains(
             response=response, text=self.test_document.label, status_code=200
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_trashed_document_list_view_with_access(self):
         self.test_document.delete()
@@ -202,15 +272,22 @@ class TrashedDocumentViewTestCase(
             obj=self.test_document, permission=permission_document_view
         )
 
+        self._clear_events()
+
         response = self._request_test_trashed_document_list_view()
         self.assertContains(
             response=response, text=self.test_document.label, status_code=200
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
     def test_trashed_document_restore_get_view_no_permission(self):
         self.test_document.delete()
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
+
+        self._clear_events()
 
         response = self._request_test_trashed_document_restore_get_view()
         self.assertEqual(response.status_code, 404)
@@ -219,6 +296,9 @@ class TrashedDocumentViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_trashed_document_restore_get_view_with_access(self):
         self.test_document.delete()
@@ -232,6 +312,8 @@ class TrashedDocumentViewTestCase(
 
         document_count = Document.valid.count()
 
+        self._clear_events()
+
         response = self._request_test_trashed_document_restore_get_view()
         self.assertEqual(response.status_code, 200)
 
@@ -240,10 +322,15 @@ class TrashedDocumentViewTestCase(
             TrashedDocument.objects.count(), trashed_document_count
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
     def test_trashed_document_restore_post_view_no_permission(self):
         self.test_document.delete()
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
+
+        self._clear_events()
 
         response = self._request_test_trashed_document_restore_post_view()
         self.assertEqual(response.status_code, 404)
@@ -252,6 +339,9 @@ class TrashedDocumentViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_trashed_document_restore_post_view_with_access(self):
         self.test_document.delete()
@@ -263,6 +353,8 @@ class TrashedDocumentViewTestCase(
             permission=permission_trashed_document_restore
         )
 
+        self._clear_events()
+
         response = self._request_test_trashed_document_restore_post_view()
         self.assertEqual(response.status_code, 302)
 
@@ -270,6 +362,14 @@ class TrashedDocumentViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count - 1
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_document)
+        self.assertEqual(events[0].verb, event_trashed_document_restored.id)
 
 
 class TrashCanViewTestCase(
@@ -287,6 +387,8 @@ class TrashCanViewTestCase(
         document_count = Document.valid.count()
         trashed_document_count = TrashedDocument.objects.count()
 
+        self._clear_events()
+
         response = self._request_empty_trash_view()
         self.assertEqual(response.status_code, 403)
 
@@ -294,6 +396,9 @@ class TrashCanViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_trash_can_empty_view_with_permission(self):
         self.test_document.delete()
@@ -303,6 +408,8 @@ class TrashCanViewTestCase(
 
         self.grant_permission(permission=permission_trash_empty)
 
+        self._clear_events()
+
         response = self._request_empty_trash_view()
         self.assertEqual(response.status_code, 302)
 
@@ -310,3 +417,11 @@ class TrashCanViewTestCase(
         self.assertEqual(
             TrashedDocument.objects.count(), trashed_document_count - 1
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_document_type)
+        self.assertEqual(events[0].verb, event_trashed_document_deleted.id)